Get rid of Step 1
Step 1 didn't add much value, fold it into Step 2.
