+++ lib/rubocop/cop/style/step1/haiku_comments.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+require 'pry'
+
+module RuboCop
+  module Cop
+    module Style
+      # Custom Cop to enforce that all comments are in haiku format
+      class HaikuComments < RuboCop::Cop::Cop
+        # TODO: Print out the comments and exit.
+      end
+    end
+  end
+end
