+++ lib/rubocop/cop/style/step5/haiku_comments.rb
@@ -8,7 +8,7 @@ module RuboCop
     module Style
       # Custom Cop to enforce that all comments are in haiku format
       class HaikuComments < RuboCop::Cop::Cop
-        # TODO: define a node matcher that calls #poetic?
+        def_node_matcher :poetic_method?, '(def _method_name _method_args `(send _receiver :puts #poetic?))'
 
         # on_new_investigation will run at the beginning of our lint process
         def on_new_investigation
@@ -23,36 +23,33 @@ module RuboCop
           node_with_comments = processed_source.ast_with_comments.select { |source_node, _comment| source_node == node }
           node = node_with_comments.keys.first
 
-          # There wasn't a matching node in the ast_with_comments
-          return if node.nil?
+          return if node.nil? # There wasn't a matching node in the ast_with_comments
 
-          # TODO: skip this node if our node matcher
           return unless poetic_method? node
 
           node_name = node.method_name.to_s
           comments = node_with_comments.values.first
 
-          return add_offense(node, message: message_for(node_name, comments)) unless comments.count == 3
+          add_offense(node, message: message_for(node_name, comments)) unless comments.count == 3
 
-          add_syllable_offenses(node_name, comments)
+          add_syllable_offenses(node, comments)
         end
 
         private
 
         # Is a method poetic, and therefore should have haiku comments?
         # This predicate method can be called inside the node_pattern DSL.
-        # A "poetic" method is one that includes 'puts' with an argument including the string 'poet'
         def poetic?(arg)
           arg.to_s.include? 'poet'
         end
 
-        def add_syllable_offenses(node_name, comments)
+        def add_syllable_offenses(node, comments)
           syllables = []
           comments.each do |comment|
             syllables << syllables(comment&.text)
           end
 
-          add_offense(node, message: message_for(node_name, comments)) unless syllables == [5, 7, 5]
+          add_offense(node, message: message_for(node.method_name.to_s, comments)) unless syllables == [5, 7, 5]
         end
 
         # Print a message that helps the developer fix their code
