--- lib/rubocop/cop/style/step6/haiku_comments.rb
@@ -1,83 +1,0 @@
-# frozen_string_literal: true
-
-require 'odyssey'
-require 'pry'
-
-module RuboCop
-  module Cop
-    module Style
-      # Custom Cop to enforce that all comments are in haiku format
-      class HaikuComments < RuboCop::Cop::Cop
-        def_node_matcher :poetic_method?, '(def _method_name _method_args `(send _receiver :puts #poetic?))'
-
-        # on_new_investigation will run at the beginning of our lint process
-        def on_new_investigation
-          define_syllable_counts
-        end
-
-        # Only runs on method definitions
-        def on_def(node)
-          # Commented source is a hash like:
-          # { <AST> => [<Parser::Source::Comment>]}
-          # Importantly, it will _only_ include the portion of the AST that has comments
-          node_with_comments = processed_source.ast_with_comments.select { |source_node, _comment| source_node == node }
-          node = node_with_comments.keys.first
-
-          return if node.nil? # There wasn't a matching node in the ast_with_comments
-
-          return unless poetic_method? node
-
-          node_name = node.method_name.to_s
-          comments = node_with_comments.values.first
-
-          add_offense(node, message: message_for(node_name, comments)) unless comments.count == 3
-
-          add_syllable_offenses(node, comments)
-        end
-
-        private
-
-        # Is a method poetic, and therefore should have haiku comments?
-        # This predicate method can be called inside the node_pattern DSL.
-        def poetic?(arg)
-          arg.to_s.include? 'poet'
-        end
-
-        def add_syllable_offenses(node, comments)
-          syllables = []
-          comments.each do |comment|
-            syllables << syllables(comment&.text)
-          end
-
-          add_offense(node, message: message_for(node.method_name.to_s, comments)) unless syllables == [5, 7, 5]
-        end
-
-        # Print a message that helps the developer fix their code
-        def message_for(node_name, comments)
-          comment_string = comments.reduce('') { |string, comment| "#{string}#{comment.text}\n" }
-          <<~MESSAGE
-            Comments for method `#{node_name}` must be in the form of a haiku.
-            Please rewrite the following comments to be more poetic:
-            #{comment_string}
-          MESSAGE
-        end
-
-        # Clean up a comment string to remove non-word characters like spaces and '#'.
-        def clean_text(text)
-          text.gsub(/\W\s/, '')
-        end
-
-        # Return the count of syllables in a string.
-        def syllables(text)
-          Odyssey.flesch_kincaid_re(clean_text(text).downcase, all_stats: true)['syllable_count']
-        end
-
-        # Odyssey doesn't always calculate syllables correctly, so we can
-        # hardcode in some words if needed, just as a workaround for the workshop.
-        def define_syllable_counts
-          Odyssey::Engine::PROBLEM_WORDS.merge!({ 'doing' => 2 })
-        end
-      end
-    end
-  end
-end
